rename to analyze
